+++ main.py
@@ -10,6 +10,12 @@ from utils.file_process import File_process
 """支援中文的"""
 embedding_model_list = ['BAAI/bge-small-zh-v1.5', 'sentence-transformers/all-MiniLM-L6-v2', 'sentence-transformers/paraphrase-multilingual-MiniLM-L12-v2', 'intfloat/multilingual-e5-large']
 
+def advanced_checkbox_change(advanced_checkbox):
+    if advanced_checkbox:
+        yield gr.update(visible=True)
+    else:
+        yield gr.update(visible=False)
+
 with gr.Blocks() as demo:
     with gr.Row():
         gr.Markdown("## Compal RAG")
@@ -22,35 +28,28 @@ with gr.Blocks() as demo:
                     no_btn = gr.Button(value="No")
                         
             upload = gr.Files(label="Upload file")
-            
-            with gr.Row():
-                text_model_dropdown = gr.Dropdown(
-                    label="Language Model", choices=LLM.get_model_list(), value=lambda: LLM.get_model(), interactive=True)
-            
-            with gr.Row():
-                load_btn = gr.Button(value="Load model")
-                unload_btn = gr.Button(value="Unload model")
-                refresh_btn = gr.Button(value="", icon="./icons/refresh.png", interactive=True)
-            
-            embed_model_dropdown = gr.Dropdown(label="Embedding Model", choices=embedding_model_list, interactive=True)
             msg_box = gr.Textbox(label="輸入問題")
-            topk = gr.Textbox(label="Top-K", value="5")
+            advanced_checkbox = gr.Checkbox(label="Advanced Options", value=False)
             submit_btn = gr.Button(value="Submit", interactive=True)
             clear_button = gr.ClearButton()
+        
+            with gr.Column(visible=False) as advanced_block:
+                embed_model_dropdown = gr.Dropdown(label="Embedding Model", choices=embedding_model_list, interactive=True) 
+                text_model_dropdown = gr.Dropdown(
+                    label="Language Model", choices=LLM.get_model_list(), value=lambda: LLM.get_model(), interactive=True)
+                topk = gr.Textbox(label="Top-K", value="5")
             status = gr.Markdown(value="")
             
         detail_output_box = gr.Chatbot(label="Detail output:", height=800)
         summary_output_box = gr.Chatbot(label="Summary output:", height=800)
     
     submit_btn.click(LLM.send_query, inputs=[text_model_dropdown, msg_box, detail_output_box, summary_output_box, topk], outputs=[msg_box, detail_output_box, summary_output_box, resp_row])
-    load_btn.click(LLM.load_model, inputs=[text_model_dropdown], outputs=[status])
-    unload_btn.click(LLM.unload_model, outputs=[status])
     upload.upload(File_process.load_file, inputs=[upload], outputs=[status])
     clear_button.add([detail_output_box, summary_output_box, msg_box])
-    refresh_btn.click(LLM.update_model_list, outputs=[text_model_dropdown])
     yes_btn.click(File_process.save_answer, inputs=[yes_btn, text_model_dropdown,  msg_box, detail_output_box, summary_output_box], outputs=[resp_row, status])
     no_btn.click(File_process.save_answer, inputs=[no_btn, text_model_dropdown,  msg_box, detail_output_box, summary_output_box], outputs=[resp_row, msg_box, detail_output_box, summary_output_box, status])
     embed_model_dropdown.change(Qdrant.load_embed_model, inputs=[embed_model_dropdown], outputs=[status])
+    advanced_checkbox.change(advanced_checkbox_change, inputs=[advanced_checkbox], outputs=[advanced_block])
 
 
 if __name__ == "__main__":
